workaround to fix hang at startup when AVG active
This is only a workaround, the core problem is still unknown.
The problem is that, when active, AVG injects its own hook (avghookx.dll or
 avghooka.dll). This, in combination with MOs own hooks cause an infinite wait
when MO tries to initialize a thread-local variable. Apparently this problem
is limited to the main thread.
The possible workarounds to prevent the hangs are: disable AVG, disable MOs hooks
for NtOpenFile, NtCreateFile AND NtClose or initialize the tls before setting
up the hooks.
The latter is what has been implemented here.