Thauthau Voodoo Fix
Issue of injected processes crashing consistently was observed by
a specific user. Crash dumps showed that VEHandler code is crashing
in the TrampolinePool::instance() in the compiler generated code
for ensuring the initialization of the static TrampolinePool.
This crash caused VEHandler to be called again and so on (not
sure how we generated the dump).
Suprisingly removing the offending code not only solved the
crashing of the VEHandler but resolved the entire issue.
Given the delicate of the TrampolinePool::instance() calls
the generate code should as simple as possible.