GetModuleFileNameW hook: fix for cases where buffer size is not large enough
Problem originally observed when path to MO is exactly 63 chars and when usvfs itself
is calling GetModuleFileNameW with a 64 byte buffer the previous code would fail (and
"reroute" the directory, returning a value of 63 indicating the buffer was large enough).

This fix aims to be more general and hopefully handle all cases properly.